requirements.txt
GitPython==2.1.11
numpy==1.16.4
tabulate==0.8.3
